python/pyproject.toml
[project]
name = "method-sdk-python-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["methodsdk>=0.0.9", "python-dotenv>=1.1.1"]
